Resolve GGQ-20251104 "Feature/task szhdz "

See merge request s25029301301/onebase-v3-be!403

+++ onebase-module-system/onebase-module-system-core/src/main/java/com/cmsr/onebase/module/system/vo/corp/CorpPageReqVO.java
@@ -21,7 +21,7 @@ public class CorpPageReqVO extends PageParam {
     private String corpName;
 
     @Schema(description = "行业类型")
-    private Integer industryType;
+    private Long industryType;
 
     @Schema(description = "状态")
     private Integer status;
